// InventarioWebApp/src/app/core/models/CodigoMedico/sarht100Model.ts
export interface sarht100Model {
    codigo: number;
    rut: string;
    dv: string;
    nombres: string;
    nombre2: string;
    apellpat: string;
    apellmat: string;
    categoriaId: number;
    servicioId: number;
    contratacionId: number;
    empresa: number;
    fechaincio: string;
    fechatermino: string;
    estado: number;
    usring: string;
    fecing: string;
    usrmod: string;
    fecmod: string;
    usr_val: number;
    fec_val: string;
    fechaincioTxt: string;
    fechaterminoTxt: string;
    idcertificadoTitulo: number | null;
    certificadoTitulo: string;
    obsTitulo: string;
    vbcertificadoTitulo: number;
    idsuperIntendencia: number | null;
    superIntendencia: string;
    obsSuperIntendencia: string;
    vbsuperIntendencia: number;
    ideunacom: number | null;
    eunacom: string;
    obsEunacom: string;
    vbeunacom: number;
    idcertificadoReconocimiento: number | null;
    certificadoReconocimiento: string;
    obsReconocimiento: string;
    vbcertificadoReconocimiento: number;
    categoriaText: string;
    servicioText: string;
    contratacionText: string;
    usuarioIng: string;

    
    colorEstado:    string;
    tipoEstado:     string;
}

// export interface sarht100Model {
//     codigo: number;
//     rut: string;
//     dv: string;
//     nombres: string;
//     nombre2: string;
//     apellpat: string;
//     apellmat: string;
//     categoriaId: number;
//     servicioId: number;
//     contratacionId: number;
//     empresa: number;
//     fechaincio?: string;
//     fechatermino?: string;
//     estado: number;
//     usring: string;
//     fecing: string;
//     usrmod: string;
//     fecmod: string;
//     usr_val: number;
//     fec_val: string;
//     fechaincioTxt: string;
//     fechaterminoTxt: string;
//     IdcertificadoTitulo?: number;
//     certificadoTitulo: string;
//     IdsuperIntendencia?: number;
//     superIntendencia: string;
//     Ideunacom?: number;
//     eunacom: string;
//     IdcertificadoReconocimiento?: number;
//     certificadoReconocimiento: string;

//     categoriaText: string;
//     servicioText: string;
//     contratacionText: string;
//     usuarioIng: string;
// }